requirements.txt
aiogram>=3.8.0
SQLAlchemy>=2.0.30
aiosqlite>=0.20.0
python-dotenv>=1.0.1
APScheduler>=3.10.4
openai>=1.40.0
pydantic>=2.8.2
pydantic-settings>=2.4.0
fastapi>=0.111.0
uvicorn>=0.30.0
